Extract type assertions into separate static helper class
Note that this removes a "feature" of the SecureHeadersTypeError
class that was able to send headers when being serialized as string.

I disagree with Aidan about the best way forward, but for the sake of
separation of concerns, this feature needs to be removed here. He will
have to decide whether it should be added again on top of these classes.

See the discussion starting in
https://github.com/aidantwoods/SecureHeaders/pull/17#issuecomment-272705873.