Pass HTTP adapter to apply() method
This makes the configuration object completely reusable.
Injecting the adapter was a (my) bad idea anyway, since
an adapter is not a dependency of the configuration class.

Now, an adapter simply needs to be passed to the `apply()`
method - calling the method without arguments will result
in the GlobalAdapter class being used: the old behavior.