add `'report-sample'` as a source shorthand ("friendly source").
Should be noted this is not adding support for the keyword (it
was always supported – keywords are not restricted. Rather, it
implements a shorthand (without the single quotes) to be used.

Info: https://www.chromestatus.com/feature/5792234276388864
Docs: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/script-src#Sources